+++ src/config.rs
@@ -362,33 +362,6 @@ impl Config {
     }
 
     #[doc(hidden)]
-    #[deprecated(
-        since = "0.31.0",
-        note = "this does nothing for now. maybe it will come back in the future."
-    )]
-    pub const fn segment_cleanup_skew(self, _: usize) -> Self {
-        self
-    }
-
-    #[doc(hidden)]
-    #[deprecated(
-        since = "0.31.0",
-        note = "this does nothing for now. maybe it will come back in the future."
-    )]
-    pub const fn segment_cleanup_threshold(self, _: u8) -> Self {
-        self
-    }
-
-    #[doc(hidden)]
-    #[deprecated(
-        since = "0.31.0",
-        note = "this does nothing for now. maybe it will come back in the future."
-    )]
-    pub fn snapshot_path<P>(self, _: P) -> Self {
-        self
-    }
-
-    #[doc(hidden)]
     pub fn flush_every_ms(mut self, every_ms: Option<u64>) -> Self {
         if Arc::strong_count(&self.0) != 1 {
             error!(